model-server/app/requirements.txt
fastapi
pydantic
requests
python-dotenv
uvicorn
langchain-google-genai
langchain
langchain-community